api: Expose Series
The main difficult is that we need to select the Series valid for the
project we're looking at ie:

  http://127.0.0.1:8000/api/1.0/projects/i915/series/

will only return the i915 series. To manage that we need to override the
get_queryset() method of the GenericModelViewSet class and properly
filter with the project name we've got from the URL.

We're also using the handy mixin classes defined by DRF to only support
a few operations on series. In particular, we don't want to be able to
create or destroy a series with the API.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>