series: Add a first attempt at a Series list view
It's all good that we can parse emails to build Series in the database,
but we also need to view them.

This commit introduces a new project/$project/series view that is the
list of series.

Instead of doing all server side, after the initial page load, we load
data through the JSON API. This should hopefully make the
editing/sorting/paging a bit less jarring as we're not reloading the
whole page anymore.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>