migrations: Use a data migration to load initial states and tags
django 1.9 is going to remove initial data through fixtures:

  If an application uses migrations, there is no automatic loading of fixtures.
  Since migrations will be required for applications in Django 1.9, this
  behavior is considered deprecated. If you want to load initial data for an
  app, consider doing it in a data migration.

  Source: https://docs.djangoproject.com/en/1.8/howto/initial-data/

So, for convenience (to not have to load fixtures by hand), load the
content of the state and tag tables as a migration step. We can then
remove a step in the install process: the initial data will populated by
the migration system (syncdb is running that under the hood).

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>