series: Log when new series are created
We create a new log entry when a new series appears.

The complexity here is that I'd like to log that a new series appeared
only when all its patches are available for a 3rd party to inspect.

That behaviour is tested, making sure an incomplete series doesn't
appear in the SeriesLog table.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>