rss: Add extra, patchwork specific, XML elements
Let's simplify the parsing work for our clients that understand
patchwork and its REST API: we can add patchwork specific XML elements,
in their own namespace, with information that will be useful to link the
RSS feed item to API objects.

Knowing which action and series correspond to the feed item are great
entry points for scripts wanting to use the RSS feed.

Reviewed-by: Damien Lespiau <damien.lespiau@intel.com>