tests: Add a utility class to create Series
So far we've been using real emails stored on-disk. While it's
interesting to have real data as input for unit tests, it's not as
flexible as having generated data. With generated data we can easily
create corner cases to improve the test coverage of the parsing code.

This is just a start, adding a TestSeries classes that can create a
series with n patches with or without a cover letter.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>